aurora-qa-system/requirements.txt
# Core API Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# HTTP Client
requests==2.31.0
httpx==0.25.1

# Data Processing
numpy==1.26.2
pandas==2.1.3

# LLM - Groq + Llama 3.1 (COMMENTED OUT)
# groq==0.11.0

# LLM - Google Gemini (COMMENTED OUT - Using Mistral)
# google-generativeai==0.3.2

# LLM - Mistral AI
mistralai==1.0.0

# Embeddings - BGE-small-en-v1.5
sentence-transformers==2.2.2
torch==2.1.0

# Search & Retrieval
rank-bm25==0.2.2

# Vector Search
faiss-cpu==1.7.4
qdrant-client==1.7.0

# NLP & Entity Extraction
spacy==3.7.2
gliner==0.1.12

# Utilities
python-dotenv==1.0.0
tqdm==4.66.1
datefinder==0.7.3
python-dateutil==2.8.2

# Development
pytest==7.4.3
black==23.11.0
